Use new CQ utility file
